refactor: consolidate game-nights/ into game-night/planning/
Move all game-nights/ planning components into game-night/planning/
subdirectory to eliminate duplicate directories while preserving
semantic separation. Update import in game night detail page.
