+++ apps/web/src/components/game-night/planning/AISuggestionCard.tsx
