+++ apps/web/src/components/game-night/planning/__tests__/AISuggestionCard.test.tsx
